+++ Client/src/Components/Organisms/Calendar/util.ts
@@ -1,7 +1,7 @@
 export const changeCalendarHeight = (current: any, height: string) => {
-  current.rootEl.current.style["overflow"] = "hidden";
-  const shit =
-    current.rootEl.current?.firstElementChild?.firstElementChild
-      ?.lastElementChild;
-  shit.style["max-height"] = height;
+  // current.rootEl.current.style["overflow"] = "hidden";
+  // const shit =
+  //   current.rootEl.current?.firstElementChild?.firstElementChild
+  //     ?.lastElementChild;
+  // shit.style["max-height"] = height;
 };
