+++ Client/src/Hook/test/index.ts
@@ -31,6 +31,7 @@ export const useChangeCalendarView = (
   const changeView = () => {
     calendarRef.current?.getInstance().changeView(view, true);
     calendarRef.current?.getInstance().prev();
+    calendarRef.current?.getInstance().render();
   };
 
   useEffect(() => {
